feat(sample): showcase issue #511 inline placeholder in tables
- Extract the ⦃check⦄ → ✅ annotator + inline-content helpers to a
  shared module file (sample/shared/.../InlineCheckPlaceholder.kt) so
  both MarkDownPage and the m3 Paparazzi Helper reuse one source.
- Wire the helpers into MarkDownPage so every sample renders the
  placeholder consistently.
- Add an "Inline placeholders (issue #511)" section to sample-tables.md
  demonstrating the substitution outside and inside table cells.
