+++ sample/shared/src/commonMain/kotlin/com/mikepenz/markdown/sample/MarkDownPage.kt
@@ -23,7 +23,6 @@ import com.mikepenz.markdown.compose.extendedspans.SquigglyUnderlineSpanPainter
 import com.mikepenz.markdown.compose.extendedspans.rememberSquigglyUnderlineAnimator
 import com.mikepenz.markdown.m3.Markdown
 import com.mikepenz.markdown.m3.elements.MarkdownCheckBox
-import com.mikepenz.markdown.model.markdownAnnotator
 import com.mikepenz.markdown.model.markdownAnnotatorConfig
 import com.mikepenz.markdown.model.markdownExtendedSpans
 import com.mikepenz.markdown.model.rememberMarkdownState
@@ -67,7 +66,8 @@ internal fun MarkDownPage(
                 },
                 checkbox = { MarkdownCheckBox(it.content, it.node, it.typography.text) },
             ),
-            annotator = markdownAnnotator(markdownAnnotatorConfig(showImageAltTooltip = true)),
+            annotator = rememberCheckAnnotator(markdownAnnotatorConfig(showImageAltTooltip = true)),
+            inlineContent = rememberCheckInlineContent(),
             imageTransformer = Coil3ImageTransformerImpl,
             extendedSpans = markdownExtendedSpans {
                 val animator = rememberSquigglyUnderlineAnimator()
